Relax reduceReducers typing to match implementation (#206)
Previously the typescript types for the `reduceReducers` function required all of the reducers to handle the initial `undefined` state, although the actual implementation ensures that only the first reducer to only ever receive an `undefined` state. The types thus forced all the reducers to implement some unnecessary handling for the initial state.

This PR allows all but the first reducer to assume that the state they receive is always initialised.